+++ lib/client/src/views/project/function-manage/api/children/create-api-sideslider/basic.vue
@@ -139,7 +139,7 @@
             const handleMethodChange = (method) => {
                 emit('update', {
                     method,
-                    query: {},
+                    query: [],
                     body: {}
                 })
             }
