+++ lib/client/src/views/project/function-manage/api/children/create-api-sideslider/edit-dynamic-data/children/dyncmic-header.tsx
@@ -1,0 +1,16 @@
+import './edit-dynamic.css'
+
+export default () => {
+    return (
+        <section class="object-layout">
+            <span class="layout-icon"></span>
+            <span class="layout-flex">属性名 (必填)</span>
+            <span class="layout-small">是否必填</span>
+            <span class="layout-item">类型</span>
+            <span class="layout-item">默认值</span>
+            <span class="layout-item">备注</span>
+            <span class="layout-icon"></span>
+            <span class="layout-icon"></span>
+        </section>
+    )
+}
