+++ lib/client/src/views/project/function-manage/api/children/create-api-sideslider/edit-dynamic-data/children/edit-dynamic.css
@@ -1,0 +1,46 @@
+.object-layout {
+    margin-top: 12px;
+    display: flex;
+    font-size: 12px;
+    .layout-flex {
+        flex: 1;
+        margin-right: 10px;
+    }
+    .layout-item {
+        width: 200px;
+        margin-right: 10px;
+    }
+    .layout-small {
+        min-width: 80px;
+        margin-right: 10px;
+    }
+    .layout-icon {
+        color: #979BA5;
+        line-height: 32px;
+        min-width: 20px;
+        cursor: pointer;
+        &.close {
+            transform: rotate(-90deg);
+        }
+        &.hidden {
+            visibility: hidden;
+        }
+        &.icon-plus-circle, &.icon-minus-circle {
+            font-size: 14px;
+        }
+        &.icon-down-shape {
+            margin-right: 4px;
+        }
+    }
+}
+.pl20 {
+    padding-left: 20px;
+}
+.property-icon {
+    display: block;
+    font-size: 12px;
+    color: #63656e;
+    &:hover {
+        color: #3a84ff;
+    }
+}
