+++ lib/client/src/views/project/function-manage/api/children/create-api-sideslider/edit-dynamic-data/index.vue
@@ -5,6 +5,7 @@
             type="unborder-card"
             :active.sync="activeTab"
             :label-height="42"
+            @tab-change="calcRenderComponent"
         >
             <bk-tab-panel
                 v-for="(panel, index) in panels"
@@ -14,9 +15,8 @@
         </bk-tab>
         <component
             class="tab-component"
-            :is="activeTab"
-            :scheme="editScheme"
-            :disable-edit-root="disableEditRoot"
+            :is="renderComponent"
+            :edit-scheme="editScheme"
             @change="handleChange"
         />
     </section>
@@ -29,37 +29,50 @@
         watch
     } from '@vue/composition-api'
     import {
-        getDefaultApiParamEditScheme,
-        API_PARAM_TYPES,
-        parseObjectValue2EditScheme,
+        parseJsonScheme2EditScheme,
         parseEditScheme2JsonScheme
     } from 'shared/api'
-    import DynamicObject from './dynamic-object.vue'
-    import PreviewObject from './preview-object.vue'
+    import DynamicObject from './children/dynamic-object.tsx'
+    import DynamicArray from './children/dynamic-array.tsx'
+    import PreviewObject from './children/preview-object.tsx'
 
     export default defineComponent({
         components: {
             DynamicObject,
+            DynamicArray,
             PreviewObject
         },
 
         props: {
-            param: Object,
-            disableEditRoot: Boolean
+            param: [Object, Array]
         },
 
         setup (props, { emit }) {
             const panels = [
-                { name: 'dynamic-object', label: '模板' },
-                { name: 'preview-object', label: '预览' }
+                { name: 'edit', label: '模板' },
+                { name: 'preview', label: '预览' }
             ]
-            const activeTab = ref('dynamic-object')
-            const editScheme = ref({})
+            const activeTab = ref('edit')
+            const renderComponent = ref(null)
+            const editScheme = ref()
+
+            const calcRenderComponent = () => {
+                if (activeTab.value === 'edit') {
+                    renderComponent.value = Array.isArray(props.param) ? DynamicArray : DynamicObject
+                } else {
+                    renderComponent.value = PreviewObject
+                }
+            }
 
             const validate = () => {
                 return new Promise((resolve, reject) => {
                     try {
-                        const jsonScheme = parseEditScheme2JsonScheme(editScheme.value)
+                        let jsonScheme
+                        if (Array.isArray(editScheme.value)) {
+                            jsonScheme = editScheme.value.map(parseEditScheme2JsonScheme)
+                        } else {
+                            jsonScheme = parseEditScheme2JsonScheme(editScheme.value)
+                        }
                         resolve(jsonScheme)
                     } catch (err) {
                         reject(err.message || err)
@@ -77,11 +90,8 @@
             watch(
                 () => props.param,
                 () => {
-                    const scheme = getDefaultApiParamEditScheme({
-                        name: 'root',
-                        type: API_PARAM_TYPES.OBJECT.VAL
-                    })
-                    editScheme.value = parseObjectValue2EditScheme(props.param, scheme)
+                    editScheme.value = parseJsonScheme2EditScheme(props.param)
+                    calcRenderComponent()
                 },
                 {
                     immediate: true
@@ -91,8 +101,10 @@
             return {
                 panels,
                 activeTab,
+                renderComponent,
                 editScheme,
                 validate,
+                calcRenderComponent,
                 handleChange
             }
         }
