+++ lib/shared/api/helper.js
@@ -21,7 +21,7 @@ export const getDefaultApi = (options = {}) => {
         versionId: undefined,
         summary: '',
         header: {},
-        query: {},
+        query: [],
         body: {},
         response: {},
         ...options
@@ -49,41 +49,54 @@ export const getDefaultApiParamEditScheme = (options = {}) => {
 
 /**
  * 数据值转换为编辑态需要的scheme
+ * 包含一些内置的逻辑，比如 root 和数组的 []
  */
-export const parseObjectValue2EditScheme = (json, scheme) => {
-    // 根据类型拼装模型
-    const type = Object.prototype.toString.call(json)
-    if (API_PARAM_TYPES.OBJECT.TYPE === type) {
-        Object.keys(json).forEach((jsonKey) => {
-            const val = json[jsonKey]
-            const childScheme = parseObjectValue2EditScheme(val, getDefaultApiParamEditScheme({
-                name: jsonKey,
-                parent: scheme
-            }))
-            scheme.properties.push(childScheme)
+export const parseJsonScheme2EditScheme = (data) => {
+    if (Array.isArray(data)) {
+        const defaultArrayScheme = [getDefaultApiParamEditScheme()]
+        const valueScheme = data.map((item) => {
+            return generateEditScheme(item)
         })
-    } else if (API_PARAM_TYPES.ARRAY.TYPE === type) {
-        // array 默认每一行数据模型相同
-        const val = json[0]
-        const childScheme = parseObjectValue2EditScheme(val, getDefaultApiParamEditScheme({
-            name: 'array-item',
-            description: '数组每个元素的模型',
+        return data.length <= 0 ? defaultArrayScheme : valueScheme
+    } else {
+        const objectScheme = getDefaultApiParamEditScheme({
+            name: 'root',
             type: API_PARAM_TYPES.OBJECT.VAL,
-            disable: true,
-            parent: scheme
-        }))
-        scheme.properties.push(childScheme)
+            default: API_PARAM_TYPES.OBJECT.DEFAULT,
+            disable: true
+        })
+        return data && data.name
+            ? generateEditScheme(data)
+            : objectScheme
+    }
+}
+
+/**
+ * 转换单个元素为编辑态 scheme
+ * @param {*} json 数据
+ * @param {*} scheme scheme
+ * @returns 最终scheme
+ */
+const generateEditScheme = (jsonScheme) => {
+    const scheme = getDefaultApiParamEditScheme({
+        name: jsonScheme.name,
+        required: jsonScheme.required,
+        default: jsonScheme.default,
+        description: jsonScheme.description,
+        type: jsonScheme.type
+    })
+
+    // 根据类型拼装模型
+    if (API_PARAM_TYPES.OBJECT.VAL === jsonScheme.type) {
+        Object.keys(jsonScheme.properties).forEach((propertyKey) => {
+            const val = jsonScheme.properties[propertyKey]
+            scheme.properties.push(generateEditScheme(val))
+        })
+    } else if (API_PARAM_TYPES.ARRAY.VAL === jsonScheme.type) {
+        scheme.properties.push(generateEditScheme(jsonScheme.items))
     } else {
         scheme.properties = []
     }
-    // 设置 scheme type & default
-    Object.keys(API_PARAM_TYPES).forEach((paramTypeKey) => {
-        const paramType = API_PARAM_TYPES[paramTypeKey]
-        if (paramType.TYPE === type) {
-            scheme.type = paramType.VAL
-            scheme.default = paramType.DEFAULT
-        }
-    })
     // 返回当前数据 scheme
     return scheme
 }
@@ -101,7 +114,8 @@ export const parseEditScheme2JsonScheme = (scheme) => {
         required: scheme.required,
         default: scheme.default,
         description: scheme.description,
-        type: scheme.type
+        type: scheme.type,
+        name: scheme.name
     }
 
     if (scheme.type === API_PARAM_TYPES.OBJECT.VAL) {
@@ -114,7 +128,7 @@ export const parseEditScheme2JsonScheme = (scheme) => {
     if (scheme.type === API_PARAM_TYPES.ARRAY.VAL) {
         result.items = []
         scheme.properties.forEach((propertyScheme) => {
-            result.items.push(parseEditScheme2JsonScheme(propertyScheme))
+            result.items = parseEditScheme2JsonScheme(propertyScheme)
         })
     }
